// src/models/CodeGraphEnvelope.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Supermodel
 * Code Graphing & Analysis API
 *
 * The version of the OpenAPI document: 0.6.3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CodeGraphStats } from './CodeGraphStats';
import {
    CodeGraphStatsFromJSON,
    CodeGraphStatsFromJSONTyped,
    CodeGraphStatsToJSON,
} from './CodeGraphStats';
import type { CodeGraphEnvelopeGraph } from './CodeGraphEnvelopeGraph';
import {
    CodeGraphEnvelopeGraphFromJSON,
    CodeGraphEnvelopeGraphFromJSONTyped,
    CodeGraphEnvelopeGraphToJSON,
} from './CodeGraphEnvelopeGraph';

/**
 * 
 * @export
 * @interface CodeGraphEnvelope
 */
export interface CodeGraphEnvelope {
    /**
     * 
     * @type {Date}
     * @memberof CodeGraphEnvelope
     */
    generatedAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof CodeGraphEnvelope
     */
    message?: string;
    /**
     * 
     * @type {CodeGraphStats}
     * @memberof CodeGraphEnvelope
     */
    stats?: CodeGraphStats;
    /**
     * 
     * @type {CodeGraphEnvelopeGraph}
     * @memberof CodeGraphEnvelope
     */
    graph: CodeGraphEnvelopeGraph;
}

/**
 * Check if a given object implements the CodeGraphEnvelope interface.
 */
export function instanceOfCodeGraphEnvelope(value: object): value is CodeGraphEnvelope {
    if (!('graph' in value) || value['graph'] === undefined) return false;
    return true;
}

export function CodeGraphEnvelopeFromJSON(json: any): CodeGraphEnvelope {
    return CodeGraphEnvelopeFromJSONTyped(json, false);
}

export function CodeGraphEnvelopeFromJSONTyped(json: any, ignoreDiscriminator: boolean): CodeGraphEnvelope {
    if (json == null) {
        return json;
    }
    return {
        
        'generatedAt': json['generatedAt'] == null ? undefined : (new Date(json['generatedAt'])),
        'message': json['message'] == null ? undefined : json['message'],
        'stats': json['stats'] == null ? undefined : CodeGraphStatsFromJSON(json['stats']),
        'graph': CodeGraphEnvelopeGraphFromJSON(json['graph']),
    };
}

export function CodeGraphEnvelopeToJSON(value?: CodeGraphEnvelope | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'generatedAt': value['generatedAt'] == null ? undefined : ((value['generatedAt']).toISOString()),
        'message': value['message'],
        'stats': CodeGraphStatsToJSON(value['stats']),
        'graph': CodeGraphEnvelopeGraphToJSON(value['graph']),
    };
}